refactoring: signing macros and controller answer spec
