+++ spec/support/feature/signing/sign_in_macros.rb
@@ -1,0 +1,10 @@
+module Feature
+  module Signing
+    def sign_in(user)
+      visit '/sign_in'
+      fill_in 'Email', with: user.email
+      fill_in 'Password', with: user.password
+      click_on 'Sign in'
+    end
+  end
+end
